+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+- Support for loading user templates
+
 ## [0.1.1] - 2025-09-06
 
 ### Changed